==> notebooks/02_transform_raw_2_ts_data.ipynb <==
import pandas as pd

rides = pd.read_parquet('../data/processed/validated_rides_2022_01.parquet')
rides.head(10)

rides['pickup_hr'] = rides['pickup_datetime'].dt.floor('H')
rides

aggregrate_rides = rides.groupby(['pickup_hr', 'pickup_loc_id']).size().reset_index()
aggregrate_rides.rename(columns={0:'rides'}, inplace=True)
aggregrate_rides

from tqdm import tqdm

def addMissingSlots(aggregrate_rides: pd.DataFrame) -> pd.DataFrame:
    
    location_ids = aggregrate_rides['pickup_loc_id'].unique()
    full_range = pd.date_range(
        aggregrate_rides['pickup_hr'].min(),
        aggregrate_rides['pickup_hr'].max(),
        freq='H' 
    )
    output = pd.DataFrame()
    
    for location_id in tqdm(location_ids):
        
        # Keep rides for specific id
        aggregrate_rides_i = aggregrate_rides.loc[aggregrate_rides.pickup_loc_id == location_id, ['pickup_hr', 'rides']]
        
        
        aggregrate_rides_i.set_index('pickup_hr', inplace=True)
        aggregrate_rides_i.index = pd.DatetimeIndex(aggregrate_rides_i.index)
        aggregrate_rides_i = aggregrate_rides_i.reindex(full_range, fill_value=0)
        
        aggregrate_rides_i['pickup_loc_id'] = location_id
        
        output = pd.concat([output, aggregrate_rides_i])
    
    output = output.reset_index().rename(columns={'index': 'pickup_hr'})
    
    return output
        

agg_rides_all_slots = addMissingSlots(aggregrate_rides)

agg_rides_all_slots

from typing import Optional, List
import plotly.express as px

def plot_rides(
    rides: pd.DataFrame,
    locations: Optional[List[int]] = None
):
    rides_2_plot = rides[rides.pickup_loc_id.isin(locations)] if locations else rides
    
    fig = px.line(
        rides_2_plot,
        x="pickup_hr",
        y="rides",
        color = 'pickup_loc_id',
        template='none',
    )
    
    fig.show()


plot_rides(agg_rides_all_slots, locations=[43])

agg_rides_all_slots.to_parquet('../data/processed/validated_rides_ts_2022_01.parquet')
