==> notebooks/04_raw_data_2_dataset.ipynb <==
%reload_ext autoreload
%autoreload 2

from src.paths import PARENT_DIR
from src.data import loadRawData, processRawData, processData2FeatTgt
rides = loadRawData(year=2022)
rides

ts_data = processRawData(rides)
ts_data.head(10)

features, targets = processData2FeatTgt(
    ts_data,
    input_feat_len= 24*28*1, # One Month
    step_size=24
)

print(f'{features.shape=}')
print(f'{targets.shape=}')
features, targets

tabular_data = features
tabular_data['tgt_rides_next_hr'] = targets

from src.paths import PROCESSED_DATA_DIR

tabular_data.to_parquet(PROCESSED_DATA_DIR / "tabular_data.parquet")
