==> notebooks/09_lightgbm_model_w_feat_eng.ipynb <==
import pandas as pd
import numpy as np
import lightgbm as lgb

from datetime import datetime
from sklearn.metrics import mean_absolute_error

from src.paths import PROCESSED_DATA_DIR
from src.data_split  import trainTestSplit

df = pd.read_parquet(PROCESSED_DATA_DIR / "tabular_data.parquet")

X_train, y_train, X_test, y_test = trainTestSplit(
    df,
    cutoff_date=datetime(2022,8,1,0,0,0),
    tgt_col_name='tgt_rides_next_hr'
)

print(f"{X_train.shape=}")
print(f"{y_train.shape=}")
print(f"{X_test.shape=}")
print(f"{y_test.shape=}")

def avgRidesPerMonth(X: pd.DataFrame) -> pd.DataFrame:
    """Adds one column with average rides from weekly data over a
     month

    Args:
        X (pd.DataFrame): _description_

    Returns:
        pd.DataFrame: _description_
    """
    X["avg_rides_per_month"] = (X[f'rides_prev_{1*7*24}_hr'] + \
                                X[f'rides_prev_{2*7*24}_hr'] + \
                                X[f'rides_prev_{3*7*24}_hr'] + \
                                X[f'rides_prev_{4*7*24}_hr']) * 0.25

    return X

from sklearn.preprocessing import FunctionTransformer

add_feat_avg_rides_month = FunctionTransformer(
    avgRidesPerMonth, validate=False
)

add_feat_avg_rides_month.fit_transform(X_train)

from sklearn.base import BaseEstimator, TransformerMixin

# Another way to add a transformation
class TemporalFeatureEngineer(BaseEstimator, TransformerMixin):
    
    def fit(self, X, y=None):
        return self
    
    def transform(self, X, y=None):
        X_ = X.copy()
        
        # Add numerical columns from datetime column
        X_['hour'] = X_['pickup_hr'].dt.hour
        X_['day_of_week'] = X_['pickup_hr'].dt.weekday
           
        return X_.drop(columns=['pickup_hr'])
        

add_temporal_feats = TemporalFeatureEngineer()
add_temporal_feats.fit_transform(X_train)

from sklearn.pipeline import make_pipeline

pipeline = make_pipeline(
    add_feat_avg_rides_month,
    add_temporal_feats,
    lgb.LGBMRegressor()
)

pipeline.fit(X_train, y_train)

predictions = pipeline.predict(X_test)

test_mae = mean_absolute_error(y_test, predictions)

print(f"Test MAE: {test_mae:.4f}")
