==> notebooks/10_lightgbm_model_w_hyperparam_tuning.ipynb <==
%reload_ext autoreload
%autoreload 2

import warnings
warnings.filterwarnings("ignore")

import pandas as pd
import numpy as np
import lightgbm as lgb

from datetime import datetime
from sklearn.pipeline import Pipeline, make_pipeline
from sklearn.metrics import mean_absolute_error
from sklearn.model_selection import KFold, TimeSeriesSplit
import optuna

from src.paths import PROCESSED_DATA_DIR
from src.data_split  import trainTestSplit
from src.model import getPipeline

df = pd.read_parquet(PROCESSED_DATA_DIR / "tabular_data.parquet")

X_train, y_train, X_test, y_test = trainTestSplit(
    df,
    cutoff_date=datetime(2022,8,1,0,0,0),
    tgt_col_name='tgt_rides_next_hr'
)

print(f"{X_train.shape=}")
print(f"{y_train.shape=}")
print(f"{X_test.shape=}")
print(f"{y_test.shape=}")

y_train

def objective(trial: optuna.trial.Trial) ->float:
    """Given set of hyper-parameters, train a model and computes an avg validation error
    based on a TimeSeriesSplit

    Args:
        trial (optuna.trial.Trial): _description_

    Returns:
        float: _description_
    """
    # Pick hyper-params
    hyperparams = {
        "metric": "mae",
        "verbose": -1,
        "num_leaves": trial.suggest_int("num_leaves", 2, 256),
        "feature_fraction": trial.suggest_float("feature_fraction", 0.2, 1.0),
        "bagging_fraction": trial.suggest_float("bagging_fraction", 0.2, 1.0),
        "min_child_samples": trial.suggest_int("min_child_samples", 3, 100)
    }
    
    ts_split = TimeSeriesSplit(n_splits=4)
    
    scores = []
    
    for train_index, val_index in ts_split.split(X_train):
        
        # Split data for training and validation
        X_train_, X_val_ = X_train.iloc[train_index, :], X_train.iloc[val_index, :]
        y_train_, y_val_ = y_train.iloc[train_index], y_train.iloc[val_index]
        
        # train the model
        pipeline = getPipeline(**hyperparams)
        pipeline.fit(X_train_, y_train_)
        
        # evaluate the model
        y_pred = pipeline.predict(X_val_)
        mae = mean_absolute_error(y_val_, y_pred)
        
        scores.append(mae)
        
    
    return np.array(scores).mean()
    
        

study = optuna.create_study(direction="minimize")
study.optimize(objective, n_trials=5)

best_params = study.best_params
print(f"Best params: {best_params}")

pipeline = getPipeline(**best_params)
pipeline.fit(X_train, y_train)

preds = pipeline.predict(X_test)
test_mae = mean_absolute_error(y_test, preds)
print(f"Test MAE: {test_mae}")


from src.plot import plotSample

plotSample(
    features=X_test,
    eg_id=2979,
    targets=y_test,
    predictions=pd.Series(preds)
)
