==> notebooks/11_backfill_feat_store.ipynb <==
%reload_ext autoreload
%autoreload 2

HOPSWORKS_PROJECT_NAME = "ml_prod_pipline"

import os
from dotenv import load_dotenv
from src.paths import PARENT_DIR

# load key-val pairs from .env file located in the parent directory
load_dotenv(PARENT_DIR / ".env")

HOPSWORKS_API_KEY = os.environ["HOPSWORKS_API_KEY"]

from datetime import datetime
import pandas as pd
from src.data import loadRawData

from_year = 2022
to_year = datetime.now().year

print(f"Downloading raw data from {from_year} to {to_year}")

rides = pd.DataFrame()

for year in range(from_year, to_year + 1):
    
    # Download data for the whole year
    rides_one_year = loadRawData(year)
    
    # Append Rows
    rides = pd.concat([rides, rides_one_year])
    


print(f"{len(rides)=}")

from src.data import processRawData

ts_data = processRawData(rides)

ts_data


ts_data
try:
    pd.to_datetime(ts_data['pickup_hr'], format='%d-%b-%Y', errors='raise')
    # do something
except ValueError:
    print("prob")
    pass

import hopsworks

project = hopsworks.login(
    project=HOPSWORKS_PROJECT_NAME,
    api_key_value=HOPSWORKS_API_KEY
)



feature_store = project.get_feature_store()

FEATURE_GROUP_NAME = "ts_hourly_feat_group"
FEATURE_GROUP_VERSION = 1

feature_group = feature_store.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description="Time-series data at hourly frequency",
    primary_key = ["pickup_loc_id", "pickup_hr"],
    event_time = "pickup_hr"
)

feature_group.insert(ts_data, write_options={"wait_for_job": False} )
