==> notebooks/13_model_training_pipeline.ipynb <==
%reload_ext autoreload
%autoreload 2

import src.config as config
import hopsworks

# Connect to project
project = hopsworks.login(
    project=config.HOPSWORKS_PROJECT_NAME,
    api_key_value=config.HOPSWORKS_API_KEY
)

# Connect to feature store
feature_store = project.get_feature_store()

# Connect to the feature group
feature_group = feature_store.get_feature_group(
    name=config.FEATURE_GROUP_NAME,
    version=config.FEATURE_GROUP_VERSION
)

# Create feature view
# Feature view uses only one feature group
print(config.FEATURE_GROUP_NAME, config.FEATURE_VIEW_NAME )
try:
    # Create feature view if one doesn't exist
    feature_store.create_feature_view(
        name=config.FEATURE_VIEW_NAME,
        version=config.FEATURE_VIEW_VERSION,
        query=feature_group.select_all()
    )

except:
    print('Feature view already existed. Skip creation...')


# Get feature view
feature_view = feature_store.get_feature_view(
    name=config.FEATURE_VIEW_NAME,
    version=config.FEATURE_VIEW_VERSION
)     

ts_data, _ = feature_view.training_data(
    description="Time series hourly taxi rides"
)

ts_data

from src.data import processData2FeatTgt

feats, tgts = processData2FeatTgt(
    ts_data,
    input_feat_len=24*28, # Month
    step_size=23,
)

feats_and_tgts = feats.copy()
feats_and_tgts['tgt_rides_nxt_hr'] = tgts

print(f"{feats_and_tgts.shape=}")

from datetime import date, timedelta
from pytz import timezone
import pandas as pd
from src.data_split import trainTestSplit

# Training data -> from Jan 2022 up until 2 months ago
# Test Data -> last 2 months

cutoff_date = pd.to_datetime(date.today() - timedelta(days=28*3))

print(f"{cutoff_date=}")


X_train, y_train, X_test, y_test = trainTestSplit(
    feats_and_tgts,
    cutoff_date,
    tgt_col_name="tgt_rides_nxt_hr"
)

print(f"{X_train.shape}")
print(f"{y_train.shape}")
print(f"{X_test.shape}")
print(f"{y_test.shape}")

import numpy as np
from sklearn.model_selection import KFold, TimeSeriesSplit
from sklearn.pipeline import make_pipeline
from sklearn.metrics import mean_absolute_error
import optuna

from src.model import getPipeline

def objective(trial: optuna.trial.Trial) -> float:
    """
    Given set of hyper-parameters, train a model and
    compute an average validation error based on a 
    TimeSeriesSplit 
    """
    
    # pick hyper-parameters
    hyperparams = {
        "metric" : "mae",
        "verbose": -1,
        "num_leaves": trial.suggest_int("num_leaves", 2, 256),
        "feature_fraction": trial.suggest_float("feature_fraction", 0.2, 1.0),
        "bagging_fraction": trial.suggest_float("bagging_fraction", 0.2, 1.0),
        "min_child_samples": trial.suggest_int("min_child_samples", 3, 100),
    }
    
    tss = TimeSeriesSplit(n_splits=2)
    scores = []
    
    for train_index, val_index in tss.split(X_train):
        
        # split data for training and validation
        X_train_, X_val_ = X_train.iloc[train_index,:],X_train.iloc[val_index,:]
        y_train_, y_val_ = y_train.iloc[train_index],y_train.iloc[val_index]
    
        # train the model
        pipeline = getPipeline(**hyperparams)
        pipeline.fit(X_train_, y_train_)
        
        # evaluate the model
        y_pred = pipeline.predict(X_val_)
        mae = mean_absolute_error(y_val_, y_pred)
        
        scores.append(mae)
        
    # Return mean score
    return np.array(scores).mean()

study  = optuna.create_study(direction="minimize")
study.optimize(objective, n_trials=1)

best_params = study.best_trial.params
print(f"{best_params=}")



pipeline = getPipeline(**best_params)
pipeline.fit(X_train,y_train)

predictions = pipeline.predict(X_test)
test_mae = mean_absolute_error(y_test, predictions)
print(f"{test_mae=:.4f}")

import joblib
from src.paths import MODELS_DIR

joblib.dump(pipeline, MODELS_DIR / "model.pkl")

from hsml.schema import Schema
from hsml.model_schema import ModelSchema

input_schema = Schema(X_train)
output_schema = Schema(y_train)
model_schema = ModelSchema(input_schema=input_schema, 
                           output_schema=output_schema)


model_registry = project.get_model_registry()

model = model_registry.sklearn.create_model(
    name="taxi_demand_predictor_next_hour",
    metrics={"test_mae": test_mae},
    description="LightGBM regressor with hyper-parameter tuning",
    input_example=X_train.sample(),
    model_schema=model_schema
)

model.save(MODELS_DIR / 'model.pkl')
