==> notebooks/14_inference_pipeline.ipynb <==
%load_ext autoreload
%autoreload 2

from datetime import datetime, timedelta
import pandas as pd 


current_date = pd.to_datetime(datetime.utcnow()).floor("H") - timedelta(28*7)
print(f"{current_date=}")

from src.inference import loadFeatBatchFromStore

features = loadFeatBatchFromStore(current_date)

from src.inference import (
    loadModelFromRegistry,
    getModelPredictions
)

model = loadModelFromRegistry()
predictions = getModelPredictions(model, features)

predictions['pickup_hr'] = current_date

predictions

from src.feature_store_api import getFeatureStore
import src.config as config

# Connect to feature group

feature_group = getFeatureStore().get_or_create_feature_group(
    name=config.FEATURE_GROUP_MODEL_PREDICTIONS,
    version=1,
    description="Model Generated Predictions",
    primary_key=["pickup_loc_id", "pickup_hr"],
    event_time="pickup_hr"
)

feature_group.insert(predictions, write_options={"wait_for_job": False})
